+++ packages/snaps-rpc-methods/src/restricted/getBip32Entropy.test.ts
@@ -75,7 +75,7 @@ describe('getBip32EntropyImplementation', () => {
     });
 
     messenger.registerActionHandler(
-      'KeyringController:withKeyringV2',
+      'KeyringController:withKeyringV2Unsafe',
       async (_selector, operation) =>
         operation({
           keyring: {
@@ -247,7 +247,7 @@ describe('getBip32EntropyImplementation', () => {
 
       expect(messenger.call).toHaveBeenCalledTimes(1);
       expect(messenger.call).toHaveBeenCalledWith(
-        'KeyringController:withKeyringV2',
+        'KeyringController:withKeyringV2Unsafe',
         { id: 'source-id' },
         expect.any(Function),
       );
@@ -287,7 +287,7 @@ describe('getBip32EntropyImplementation', () => {
 
       expect(messenger.call).toHaveBeenCalledTimes(1);
       expect(messenger.call).toHaveBeenCalledWith(
-        'KeyringController:withKeyringV2',
+        'KeyringController:withKeyringV2Unsafe',
         { id: 'source-id' },
         expect.any(Function),
       );
@@ -341,7 +341,7 @@ describe('getBip32EntropyImplementation', () => {
       >({ namespace: MOCK_ANY_NAMESPACE });
 
       messenger.registerActionHandler(
-        'KeyringController:withKeyringV2',
+        'KeyringController:withKeyringV2Unsafe',
         async (_selector, operation) =>
           operation({
             keyring: {
@@ -371,7 +371,7 @@ describe('getBip32EntropyImplementation', () => {
       >({ namespace: MOCK_ANY_NAMESPACE });
 
       messenger.registerActionHandler(
-        'KeyringController:withKeyringV2',
+        'KeyringController:withKeyringV2Unsafe',
         async (_selector, operation) =>
           operation({
             keyring: {
@@ -401,7 +401,7 @@ describe('getBip32EntropyImplementation', () => {
       >({ namespace: MOCK_ANY_NAMESPACE });
 
       messenger.registerActionHandler(
-        'KeyringController:withKeyringV2',
+        'KeyringController:withKeyringV2Unsafe',
         async (_selector, operation) =>
           operation({
             keyring: {
@@ -435,7 +435,7 @@ describe('getBip32EntropyImplementation', () => {
       >({ namespace: MOCK_ANY_NAMESPACE });
 
       messenger.registerActionHandler(
-        'KeyringController:withKeyringV2',
+        'KeyringController:withKeyringV2Unsafe',
         async (_selector, operation) =>
           operation({
             keyring: {