+++ packages/snaps-rpc-methods/src/restricted/getBip32PublicKey.test.ts
@@ -75,7 +75,7 @@ describe('getBip32PublicKeyImplementation', () => {
     });
 
     messenger.registerActionHandler(
-      'KeyringController:withKeyringV2',
+      'KeyringController:withKeyringV2Unsafe',
       async (_selector, operation) =>
         operation({
           keyring: {
@@ -223,7 +223,7 @@ describe('getBip32PublicKeyImplementation', () => {
 
       expect(messenger.call).toHaveBeenCalledTimes(1);
       expect(messenger.call).toHaveBeenCalledWith(
-        'KeyringController:withKeyringV2',
+        'KeyringController:withKeyringV2Unsafe',
         { id: 'source-id' },
         expect.any(Function),
       );
@@ -253,7 +253,7 @@ describe('getBip32PublicKeyImplementation', () => {
 
       expect(messenger.call).toHaveBeenCalledTimes(1);
       expect(messenger.call).toHaveBeenCalledWith(
-        'KeyringController:withKeyringV2',
+        'KeyringController:withKeyringV2Unsafe',
         { id: 'source-id' },
         expect.any(Function),
       );