+++ packages/snaps-rpc-methods/src/restricted/getBip44Entropy.test.ts
@@ -74,7 +74,7 @@ describe('getBip44EntropyImplementation', () => {
     });
 
     messenger.registerActionHandler(
-      'KeyringController:withKeyringV2',
+      'KeyringController:withKeyringV2Unsafe',
       async (_selector, operation) =>
         operation({
           keyring: {
@@ -150,7 +150,7 @@ describe('getBip44EntropyImplementation', () => {
 
       expect(messenger.call).toHaveBeenCalledTimes(1);
       expect(messenger.call).toHaveBeenCalledWith(
-        'KeyringController:withKeyringV2',
+        'KeyringController:withKeyringV2Unsafe',
         { id: 'source-id' },
         expect.any(Function),
       );